requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
transformers
openai-whisper
torch
python-multipart
